- Replace video player package by better player plus. - Update dependencies and refactor story view implementation. - Add RTL support to story.
