+++ lib/story_view.dart
@@ -1,5 +1,11 @@
-export 'widgets/story_image.dart';
-export 'widgets/story_video.dart';
+export 'widgets/image/story_image.dart';
+export 'widgets/video/story_video.dart';
 export 'widgets/story_view.dart';
 export 'controller/story_controller.dart';
-export 'utils.dart';
+export 'utils/utils.dart';
+export 'utils/contrast_helper.dart';
+export 'utils/enum.dart';
+export 'utils/video_loader.dart';
+export 'utils/page_data.dart';
+export 'widgets/page_bar.dart';
+export 'widgets/story_item.dart';
