+++ lib/utils/contrast_helper.dart
@@ -1,0 +1,20 @@
+import 'dart:math';
+
+/// Concept source: https://stackoverflow.com/a/9733420
+class ContrastHelper {
+  static double luminance(double? r, double? g, double? b) {
+    final a = [r, g, b].map((it) {
+      double value = it!.toDouble() / 255.0;
+      return value <= 0.03928
+          ? value / 12.92
+          : pow((value + 0.055) / 1.055, 2.4);
+    }).toList();
+
+    return a[0] * 0.2126 + a[1] * 0.7152 + a[2] * 0.0722;
+  }
+
+  static double contrast(rgb1, rgb2) {
+    return luminance(rgb2[0], rgb2[1], rgb2[2]) /
+        luminance(rgb1[0], rgb1[1], rgb1[2]);
+  }
+}
