+++ lib/utils/enum.dart
@@ -1,0 +1,10 @@
+enum PlaybackState { pause, play, next, previous }
+
+/// Indicates where the progress indicators should be placed.
+enum ProgressPosition { top, bottom, none }
+
+enum LoadState { loading, success, failure }
+
+enum Direction { up, down, left, right }
+
+enum StoryItemType { text, image, video }
