+++ lib/utils/page_data.dart
@@ -1,0 +1,8 @@
+/// Capsule holding the duration and shown property of each story. Passed down
+/// to the pages bar to render the page indicators.
+class PageData {
+  Duration duration;
+  bool shown;
+
+  PageData(this.duration, this.shown);
+}
