--- lib/widgets/story_video.dart
@@ -1,193 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:video_player/video_player.dart';
-
-import '../utils.dart';
-import '../controller/story_controller.dart';
-
-class VideoLoader {
-  String url;
-
-  File? videoFile;
-
-  Map<String, dynamic>? requestHeaders;
-
-  LoadState state = LoadState.loading;
-
-  VideoLoader(this.url, {this.requestHeaders});
-
-  void loadVideo(VoidCallback onComplete) {
-    if (this.videoFile != null) {
-      this.state = LoadState.success;
-      onComplete();
-    }
-
-    final fileStream = DefaultCacheManager()
-        .getFileStream(this.url, headers: this.requestHeaders as Map<String, String>?);
-
-    fileStream.listen((fileResponse) {
-      if (fileResponse is FileInfo) {
-        if (this.videoFile == null) {
-          this.state = LoadState.success;
-          this.videoFile = fileResponse.file;
-          onComplete();
-        }
-      }
-    });
-  }
-}
-
-class StoryVideo extends StatefulWidget {
-  final StoryController? storyController;
-  final VideoLoader videoLoader;
-  final Widget? loadingWidget;
-  final Widget? errorWidget;
-
-  StoryVideo(this.videoLoader, {
-    Key? key,
-    this.storyController,
-    this.loadingWidget,
-    this.errorWidget,
-  }) : super(key: key ?? UniqueKey());
-
-  static StoryVideo url(String url, {
-    StoryController? controller,
-    Map<String, dynamic>? requestHeaders,
-    Key? key,
-    Widget? loadingWidget,
-    Widget? errorWidget,
-  }) {
-    return StoryVideo(
-      VideoLoader(url, requestHeaders: requestHeaders),
-      storyController: controller,
-      key: key,
-      loadingWidget: loadingWidget,
-      errorWidget: errorWidget,
-    );
-  }
-
-  @override
-  State<StatefulWidget> createState() {
-    return StoryVideoState();
-  }
-}
-
-class StoryVideoState extends State<StoryVideo> {
-  Future<void>? playerLoader;
-
-  StreamSubscription? _streamSubscription;
-
-  VideoPlayerController? playerController;
-
-  @override
-  void initState() {
-    super.initState();
-
-    widget.storyController!.pause();
-
-    widget.videoLoader.loadVideo(() {
-      if (widget.videoLoader.state == LoadState.success) {
-        this.playerController =
-            VideoPlayerController.file(widget.videoLoader.videoFile!);
-
-        playerController!.initialize().then((v) {
-          setState(() {});
-          widget.storyController!.play();
-        });
-
-        if (widget.storyController != null) {
-          _streamSubscription =
-              widget.storyController!.playbackNotifier.listen((playbackState) {
-            if (playbackState == PlaybackState.pause) {
-              playerController!.pause();
-            } else {
-              playerController!.play();
-            }
-          });
-        }
-      } else {
-        setState(() {});
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.black,
-      height: double.infinity,
-      width: double.infinity,
-      child: VideoContentView(
-        videoLoadState: widget.videoLoader.state,
-        playerController: playerController,
-        loadingWidget: widget.loadingWidget,
-        errorWidget: widget.errorWidget,
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    playerController?.dispose();
-    _streamSubscription?.cancel();
-    super.dispose();
-  }
-}
-
-/**
- * @name VideoContentView
- * @description Stateless widget that shows a video player or loading/error widgets based on video loading state.
- */
-class VideoContentView extends StatelessWidget {
-  final LoadState videoLoadState;
-  final VideoPlayerController? playerController;
-  final Widget? loadingWidget;
-  final Widget? errorWidget;
-
-  const VideoContentView({
-    Key? key,
-    required this.videoLoadState,
-    required this.playerController,
-    this.loadingWidget,
-    this.errorWidget,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    if (videoLoadState == LoadState.success &&
-        playerController != null &&
-        playerController!.value.isInitialized) {
-      return Center(
-        child: AspectRatio(
-          aspectRatio: playerController!.value.aspectRatio,
-          child: VideoPlayer(playerController!),
-        ),
-      );
-    }
-
-    if (videoLoadState == LoadState.loading) {
-      return Center(
-        child: loadingWidget ??
-            const SizedBox(
-              width: 70,
-              height: 70,
-              child: CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                strokeWidth: 3,
-              ),
-            ),
-      );
-    }
-
-    return Center(
-      child: errorWidget ??
-          const Text(
-            "Media failed to load.",
-            style: TextStyle(color: Colors.white),
-          ),
-    );
-  }
-}
