+++ pubspec.yaml
@@ -1,15 +1,16 @@
 name: story_view
 description: Display stories just like Whatsapp & Instagram. Can be used inline/inside ListView or Column just like Google News app. Comes with gestures to pause, forward & go to previous page.
-version: 0.16.6
+version: 0.17.0
 homepage: https://github.com/blackmann/story_view
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: ^3.9.0
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_cache_manager: ^3.4.1
   rxdart: ^0.28.0
-  video_player: ^2.10.0
   collection: ^1.19.1
+  better_player_plus: ^1.0.8
+
